sched: core: fix warning
kernel/sched/core.c: In function 'idle_cpu_relaxed':
kernel/sched/core.c:6474:39: warning: comparison between pointer and integer
  if (cpu_relaxed_read_long(&rq->curr) != rq->idle)
                                       ^~

Fixes: e5434290a23c ("kernel: cpuidle: use idle_cpu_relaxed and cpu_relaxed")

+++ kernel/sched/core.c
@@ -6449,8 +6449,8 @@ int idle_cpu_relaxed(int cpu)
 {
       struct rq *rq = cpu_rq(cpu);
 
-      if (cpu_relaxed_read_long(&rq->curr) != rq->idle)
-             return 0;
+	if (cpu_relaxed_read_long(&rq->curr) != (u64)rq->idle)
+		return 0;
 
       if (cpu_relaxed_read_long(&rq->nr_running))
              return 0;
